sync upstream changes (#2)
* Bugfix for losing autoFilter with multiple sheets

* Check that tempfiles could be created.

I spend some time with random and strange php errors being
emitted like :

    tempnam(): file created in the system's temporary directory

and fopen()'s failing. This turned out to be because I was exceeding
ulimits. Checking that the file was actually created will help the
next person who hits this problem!

* Problem with row options on Excel 2007

If you specify some row options (like font-style or border), boolean value is not translated as string, so some attribut get empty value, not the "false" string

* Fix '=' in Google Sheets

* Add missing property, fix call to htmlspecialchars by ensuring the parameter is string

* adding hyperlink example

* github/issues/324 fix

* allow a call to writeRow before writeHeader

* disallowing invalid unicode in sheet name

* changing composer type to library

---------

Co-authored-by: heew <wow.howard_94@hotmail.com>
Co-authored-by: Rob Thomas <rthomas@clearlyip.com>
Co-authored-by: mk-j <mk-j@users.noreply.github.com>
Co-authored-by: Vincent <fulgurio@users.noreply.github.com>
Co-authored-by: kkr-Christopher <91697610+kkr-Christopher@users.noreply.github.com>
Co-authored-by: asalonen <arttu.salonen@hoistgroup.com>
Co-authored-by: mk-j <mark@zedwood.com>